Revert "Hardcode sample_rate for 32bit support"
This reverts commit b88bb3ba028c58fd4ddd563676131313719dbf92.

+++ src/octaver/octaver.rs
@@ -10,7 +10,7 @@ pub struct Octaver {
 }
 
 impl Octaver {
-  pub fn new(sample_rate: f32) -> Self {
+  pub fn new(sample_rate: f64) -> Self {
     Self {
       lowpass: Lowpass::new(),
       delta: Delta::new(),
